MOD #264 : 파사드 계층 구조 컨벤션 적용해서 단일 책임으로 메서드 분리
