+++ src/main/java/com/lokoko/domain/user/application/service/UserGetService.java
@@ -1,0 +1,21 @@
+package com.lokoko.domain.user.application.service;
+
+import com.lokoko.domain.user.domain.entity.User;
+import com.lokoko.domain.user.domain.repository.UserRepository;
+import com.lokoko.domain.user.exception.UserNotFoundException;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class UserGetService {
+
+    private final UserRepository userRepository;
+
+    public User findUserById(Long userId) {
+        return userRepository.findById(userId)
+                .orElseThrow(UserNotFoundException::new);
+    }
+}
