+++ src/main/java/com/lokoko/global/auth/tiktok/service/TikTokConnectionService.java
@@ -1,13 +1,12 @@
 package com.lokoko.global.auth.tiktok.service;
 
+import com.lokoko.domain.creator.application.service.CreatorSaveService;
 import com.lokoko.domain.creator.domain.entity.Creator;
-import com.lokoko.domain.creator.domain.repository.CreatorRepository;
+import com.lokoko.domain.customer.application.CustomerSaveService;
 import com.lokoko.domain.customer.domain.entity.Customer;
-import com.lokoko.domain.customer.domain.repository.CustomerRepository;
+import com.lokoko.domain.user.application.service.UserGetService;
 import com.lokoko.domain.user.domain.entity.User;
 import com.lokoko.domain.user.domain.entity.enums.Role;
-import com.lokoko.domain.user.domain.repository.UserRepository;
-import com.lokoko.domain.user.exception.UserNotFoundException;
 import com.lokoko.global.auth.exception.ErrorMessage;
 import com.lokoko.global.auth.exception.OauthException;
 import com.lokoko.global.auth.tiktok.TikTokOAuthClient;
@@ -23,24 +22,24 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @RequiredArgsConstructor
 public class TikTokConnectionService {
-    
+
     private final TikTokOAuthClient tikTokOAuthClient;
-    private final UserRepository userRepository;
-    private final CreatorRepository creatorRepository;
-    private final CustomerRepository customerRepository;
+    private final UserGetService userGetService;
+
+    private final CustomerSaveService customerSaveService;
+    private final CreatorSaveService creatorSaveService;
     private final TikTokRedisTokenService tikTokRedisTokenService;
 
     public String generateConnectionUrl(Long userId) {
 
-        userRepository.findById(userId);
+        userGetService.findUserById(userId);
         return tikTokOAuthClient.buildAuthorizationUrl(userId);
     }
 
     @Transactional
     public TikTokConnectionResponse connectTikTok(Long userId, String code) {
         try {
-            User user  = userRepository.findById(userId)
-                    .orElseThrow(UserNotFoundException::new);
+            User user = userGetService.findUserById(userId);
 
             // 토큰 발급
             TikTokTokenDto tokenDto = tikTokOAuthClient.issueToken(code);
@@ -57,15 +56,15 @@ public class TikTokConnectionService {
             // 토큰을 Redis에 저장
             saveTokensInRedis(userId, tokenDto);
 
-            if (user.getRole() == Role.CREATOR){
+            if (user.getRole() == Role.CREATOR) {
                 Creator creator = user.getCreator();
                 creator.connectTikTok(profileDto.openId());
-                creatorRepository.save(creator);
+                creatorSaveService.save(creator);
 
             } else if (user.getRole() == Role.CUSTOMER) {
                 Customer customer = user.getCustomer();
                 customer.connectTikTok(profileDto.openId());
-                customerRepository.save(customer);
+                customerSaveService.save(customer);
             }
 
             return TikTokConnectionResponse.connected(profileDto.openId());
@@ -80,10 +79,10 @@ public class TikTokConnectionService {
 
     private void saveTokensInRedis(Long userId, TikTokTokenDto tokenDto) {
         tikTokRedisTokenService.saveTokens(userId,
-            tokenDto.accessToken(),
-            tokenDto.refreshToken(),
-            tokenDto.expiresIn(),
-            tokenDto.refreshExpiresIn()
+                tokenDto.accessToken(),
+                tokenDto.refreshToken(),
+                tokenDto.expiresIn(),
+                tokenDto.refreshExpiresIn()
         );
     }
